add fuzzy matching and other suggestions
