+++ app/internal/ranking/download_ranking.py
@@ -10,7 +10,7 @@ from sqlmodel import Session
 from app.internal.models import BookRequest, ProwlarrSource
 from app.internal.ranking.quality import quality_config
 from app.internal.ranking.quality_extract import Quality, extract_qualities
-from app.internal.mam.mam import mam_config
+
 
 class RankSource(pydantic.BaseModel):
     source: ProwlarrSource
@@ -178,22 +178,30 @@ class CompareSource:
         return int(b_title) - int(a_title)
 
     def _compare_authors(self, a: RankSource, b: RankSource, next_compare: int) -> int:
-        if(mam_config.is_active(self.session) and (a.source.authors!=[] or b.source.authors!=[])):
-            a_score =  get_intersection_length(a.source.authors, self.book.authors)
-            b_score =  get_intersection_length(b.source.authors, self.book.authors)
-        else:
-
-
-            a_score = vaguely_exist_in_title(
+        a_score = max(
+            vaguely_exist_in_title(
                 self.book.authors,
                 a.source.title,
                 quality_config.get_name_exists_ratio(self.session),
-            )
-            b_score = vaguely_exist_in_title(
+            ),
+            fuzzy_author_narrator_match(
+                a.source.authors,
+                self.book.authors,
+                quality_config.get_name_exists_ratio(self.session),
+            ),
+        )
+        b_score = max(
+            vaguely_exist_in_title(
                 self.book.authors,
                 b.source.title,
                 quality_config.get_name_exists_ratio(self.session),
-            )
+            ),
+            fuzzy_author_narrator_match(
+                b.source.authors,
+                self.book.authors,
+                quality_config.get_name_exists_ratio(self.session),
+            ),
+        )
         if a_score == b_score:
             return self._get_next_compare(next_compare)(a, b, next_compare + 1)
         return b_score - a_score
@@ -201,20 +209,30 @@ class CompareSource:
     def _compare_narrators(
         self, a: RankSource, b: RankSource, next_compare: int
     ) -> int:
-        if(mam_config.is_active(self.session) and (a.source.narrators!=[] or b.source.narrators!=[])):
-            a_score =  get_intersection_length(a.source.authors, self.book.authors)
-            b_score =  get_intersection_length(b.source.authors, self.book.authors)
-        else:
-            a_score = vaguely_exist_in_title(
+        a_score = max(
+            vaguely_exist_in_title(
                 self.book.narrators,
                 a.source.title,
                 quality_config.get_name_exists_ratio(self.session),
-            )
-            b_score = vaguely_exist_in_title(
+            ),
+            fuzzy_author_narrator_match(
+                a.source.narrators,
+                self.book.narrators,
+                quality_config.get_name_exists_ratio(self.session),
+            ),
+        )
+        b_score = max(
+            vaguely_exist_in_title(
                 self.book.narrators,
                 b.source.title,
                 quality_config.get_name_exists_ratio(self.session),
-            )
+            ),
+            fuzzy_author_narrator_match(
+                b.source.narrators,
+                self.book.narrators,
+                quality_config.get_name_exists_ratio(self.session),
+            ),
+        )
         if a_score == b_score:
             return self._get_next_compare(next_compare)(a, b, next_compare + 1)
         return b_score - a_score
@@ -234,7 +252,29 @@ class CompareSource:
             return int((a.source.publish_date - b.source.publish_date).total_seconds())
         # With torrents: older => better
         return int((b.source.publish_date - a.source.publish_date).total_seconds())
-    
+
+
+def fuzzy_author_narrator_match(
+    source_people: list[str], book_people: list[str], name_exists_ratio: int
+) -> int:
+    """Calculate a fuzzy matching score between two lists of author/narrator names."""
+    if not source_people or not book_people:
+        return 0
+    score = 0
+    for book_person in book_people:
+        best_match = 0
+        for source_person in source_people:
+            match_score = fuzz.token_set_ratio(
+                book_person, source_person, processor=utils.default_process
+            )
+            best_match = max(best_match, match_score)
+
+        # Only count matches above threshold
+        if best_match > name_exists_ratio:
+            score += 1
+
+    return score
+
 
 def vaguely_exist_in_title(words: list[str], title: str, name_exists_ratio: int) -> int:
     return sum(
@@ -243,10 +283,10 @@ def vaguely_exist_in_title(words: list[str], title: str, name_exists_ratio: int)
         if fuzz.token_set_ratio(w, title, processor=utils.default_process)
         > name_exists_ratio
     )
-def get_intersection_length(a : list[str],b: list[str]):
+
+
+def get_intersection_length(a: list[str], b: list[str]):
     return len(set(a).intersection(set(b)))
-
-
 
 
 def exists_in_title(word: str, title: str, title_exists_ratio: int) -> bool:
