+++ app/routers/wishlist.py
@@ -26,9 +26,7 @@ from app.internal.prowlarr.prowlarr import (
     prowlarr_config,
     start_download,
 )
-from app.internal.mam.mam import (
-    mam_config
-)
+from app.internal.indexers.mam import mam_config
 from app.internal.query import query_sources
 from app.internal.ranking.quality import quality_config
 from app.util.auth import DetailedUser, get_authenticated_user
@@ -202,7 +200,7 @@ async def list_sources(
         {
             "book": result.book,
             "sources": result.sources,
-            "mam_active": mam_config.is_active(session)
+            "mam_active": mam_config.is_active(session),
         },
     )
